+++ src/chanjo2/crud/intervals.py
@@ -10,6 +10,7 @@ from chanjo2.models.pydantic_models import (
     ExonBase,
     GeneBase,
     TranscriptBase,
+    TranscriptTag,
 )
 from chanjo2.models.sql_models import Exon as SQLExon
 from chanjo2.models.sql_models import Gene as SQLGene
@@ -124,15 +125,27 @@ def get_hgnc_gene(db: Session, build: Builds, hgnc_id: int) -> SQLGene:
     return gene_query.first()
 
 
+def _filter_transcripts_by_tag(
+    transcripts: query.Query, transcript_tags: List[TranscriptTag] = []
+) -> query.Query:
+    """Return transcripts which contain one or more RefSeq tag."""
+    for tag in transcript_tags:
+        transcripts: query.Query = transcripts.filter(
+            getattr(SQLTranscript, tag).isnot(None)
+        )
+    return transcripts
+
+
 def get_gene_intervals(
     db: Session,
     build: Builds,
-    ensembl_ids: Optional[List[str]],
-    hgnc_ids: Optional[List[int]],
-    hgnc_symbols: Optional[List[str]],
-    ensembl_gene_ids: Optional[List[str]],
-    limit: Optional[int],
     interval_type: Union[SQLTranscript, SQLExon],
+    ensembl_ids: Optional[List[str]] = [],
+    hgnc_ids: Optional[List[int]] = [],
+    hgnc_symbols: Optional[List[str]] = [],
+    ensembl_gene_ids: Optional[List[str]] = [],
+    limit: Optional[int] = None,
+    transcript_tags: Optional[List[TranscriptTag]] = [],
 ) -> List[Union[SQLTranscript, SQLExon]]:
     """Retrieve transcripts or exons from a list of genes."""
 
@@ -149,6 +162,11 @@ def get_gene_intervals(
         intervals: query.Query = intervals.filter(SQLGene.hgnc_id.in_(hgnc_ids))
     elif hgnc_symbols:
         intervals: query.Query = intervals.filter(SQLGene.hgnc_symbol.in_(hgnc_symbols))
+
+    if interval_type == SQLTranscript and transcript_tags:
+        intervals = _filter_transcripts_by_tag(
+            transcripts=intervals, transcript_tags=transcript_tags
+        )
 
     intervals: query.Query = intervals.filter(interval_type.build == build)